+++ docs/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Fixed
+- Filter out empty events from history API resutls (OSIDB-3942)
+
 ## [4.7.0] - 2025-01-28
 ### Added
 - Add SLA exclusion policies (OSIDB-3711)